.docker/local/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["yuya0628arsaga <yuya.ogawa@arsaga.jp>"]

[tool.poetry.dependencies]
python = "3.9.13"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"

python-dotenv = "^1.0.0"
langchain = "^0.0.354"
openai = "^1.6.1"
chromadb = "0.4.17"
langchain-experimental = "^0.0.47"
tiktoken = "^0.5.2"
google-cloud-firestore = "^2.14.0"
firebase-admin = "^6.3.0"
boto3 = "^1.34.17"

pysqlite3-binary = "^0.5.2"
pysqlite3 = "^0.5.2"

# PDF処理用
"pdfminer.six" = "^20231228"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
